move CLObject implementations to their own file
