+++ pytissueoptics/rayscattering/opencl/CLObjects/__init__.py
@@ -1,3 +1,12 @@
 from .CLObject import CLObject
-from .propagation import *
-from .scene import *
+
+from .dataPointCL import DataPointCL
+from .materialCL import MaterialCL
+from .photonCL import PhotonCL
+from .seedCL import SeedCL
+
+from .solidCandidateCL import SolidCandidateCL
+from .solidCL import SolidCL, SolidCLInfo
+from .surfaceCL import SurfaceCL, SurfaceCLInfo
+from .triangleCL import TriangleCL, TriangleCLInfo
+from .vertexCL import VertexCL
