+++ pytissueoptics/rayscattering/opencl/CLScene.py
@@ -3,9 +3,14 @@ from typing import List
 import numpy as np
 
 from pytissueoptics.rayscattering.tissues import RayScatteringScene
-from pytissueoptics.rayscattering.opencl.CLObjects import MaterialCL, SolidCandidateCL, SolidCL, SolidCLInfo, \
-    SurfaceCLInfo, SurfaceCL, TriangleCLInfo, TriangleCL, VertexCL
-
+from pytissueoptics.rayscattering.opencl.CLObjects import SolidCLInfo, \
+    SurfaceCLInfo, TriangleCLInfo
+from pytissueoptics.rayscattering.opencl.CLObjects.solidCandidateCL import SolidCandidateCL
+from pytissueoptics.rayscattering.opencl.CLObjects.vertexCL import VertexCL
+from pytissueoptics.rayscattering.opencl.CLObjects.triangleCL import TriangleCL
+from pytissueoptics.rayscattering.opencl.CLObjects.surfaceCL import SurfaceCL
+from pytissueoptics.rayscattering.opencl.CLObjects.solidCL import SolidCL
+from pytissueoptics.rayscattering.opencl.CLObjects.materialCL import MaterialCL
 
 NO_LOG_ID = 0
 NO_SOLID_ID = -1
